\name{msir-package}
\alias{msir-package}
\docType{package}
\title{Model-based Sliced Inverse Regression}
\description{An R package for dimension reduction using Model-based Sliced Inverse Regression.}
\details{
\tabular{ll}{
Package: \tab msir\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-09\cr
License: \tab GPL version 2 (or newer) \cr
LazyLoad: \tab yes\cr
}
}
\author{Luca Scrucca \email{luca@stat.unipg.it}}
\references{Scrucca, L. (2011) Model-based SIR for dimension reduction, to appear in Computational Statistics & Data Analysis.}
%\seealso{\code{\link[dr:dr-package]{dr}}}
%\examples{}
\keyword{package}
